+++ lib/V8/CMakeLists.txt
@@ -24,16 +24,6 @@ target_include_directories(arango_v8
   ${OPENSSL_INCLUDE_DIR}
   ${ICU_INCLUDE_DIR})
 
-add_library(arango_v8_deadline STATIC
-  processMonitor.cpp
-  v8-deadline.cpp)
-target_link_libraries(arango_v8_deadline
-  arango
-  v8_interface
-  velocypack)
-target_include_directories(arango_v8_deadline PRIVATE
-  ${CMAKE_CURRENT_SOURCE_DIR})
-
 add_library(arango_v8_no_deadline STATIC
   v8-no-deadline.cpp)
 target_link_libraries(arango_v8_no_deadline
